simple-yfinance-agent/requirements.txt
langchain
langchain-community
langchain-openai
python-dotenv
pydantic
duckduckgo-search
langchain-core
langchain-classic
duckduckgo-search
langgraph
yfinance